Add source directory to include path
Relative includes are no longer relative to the file as they ideally
would but relative to the source directory. Add missing include
directive for builds using the kernel infrastructure to build the
module. This is the same workaround as in the default Makefile.

As the default Makfile can also be used without the wrapping target,
hence has the same issue with the relative includes, use Kbuild
facilities to set CFLAGS and tidy up the Makefile.

Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>