Add support for Linux 4.12
In preparation to add HE support Linux 4.12 separates flags for
bandwidth and encoding in mac80211. Use the new flags as appropriate.

To avoid using LINUX_VERSION_CODE, as requested, pull a stunt using the
definition of RX_ENC_FLAG_STBC_SHIFT which was added in 4.12 as well to
check for the version thereby obfuscating what is going on.

URL: https://github.com/kaloz/mwlwifi/pull/156
Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>