fix reinjected tx_skb and remove queue amsdu_pkts The amsdu->skb paquet is created, but never used.
check the code:
skb_queue_tail(amsdu_pkts, tx_skb);

if (amsdu->num > SYSADPT_AMSDU_PACKET_THRESHOLD) {
        amsdu->num = 0;
        amsdu->cur_pos = NULL;
        spin_unlock_bh(&sta_info->amsdu_lock);
        return amsdu->skb;

Even if we return amsdu->skb, it is tx_skb that is put in the amsdu_pkts tail. And since it goes through, amsdu->skb is never really processed.
Surprisingly, with this fix, we realize that the queue is created only for one package.
As a reminder, AMSDU is an aggregation of MSDUs, so it has to go through the sk_buff->data.
The passage in the current queue is sufficient.

To be studied if it is really useful. This commit fixes the problem of this driver.

Signed-off-by: Julien Blais webmaster@jbsky.fr